track timestamps and generate midi events
This stores note on/off events in an arrray.array('L') of uint32
values. The tick timestamps (units of pulses per quarter note) go
in the high bits, so it should be possible to sort all note on/off
events, from all voices, into a single ordered list of midi events.